backend_python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1
google-genai
pypdf==4.0.1
python-docx==1.1.0
requests==2.31.0
passlib[argon2]
python-jose[cryptography]
python-multipart
argon2-cffi
